+++ shout-fip.py
@@ -22,6 +22,9 @@ def killbuffer(signum, frame):  # pylint: disable=unused-argument
 def cleantmpdir(tmpdir):
     n = 0
     for tmpfn in os.listdir(tmpdir):
+        if os.path.isdir(tmpfn):
+            sys.stdout.write("\nNot removing directory %s " % os.path.join(tmpdir,tmpfn))
+            continue
         sys.stdout.write("\rClearning %s " % os.path.join(tmpdir,tmpfn))
         sys.stdout.flush()
         os.remove(os.path.join(tmpdir,tmpfn))
